split resolvers.ts into theme.ts and language.ts
